requirements.txt
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
